feat(about/): added new page Added a new about page. Split the main page and the about me page. This is so that can add more information and design instead of cramming everything into the main page.
refactor(navbar.tsx, page.tsx): Made logo clickable and added new tab
Made the logo clickable to link back the the main page. Added the
about tab to navigate to the about me page. Also changed the dimensions of
the image on the home page to make it not so large.