feat(api/data/route.tsx, skills.tsx, interfaces/): Added dynamic loading of skills and projects
The data api folder will handle the loading of the json data dynamically.
The skills.tsx handles the the skills cards section as a component to
split up the work from the main `myWork` page.

The interfaces just holds the types for each skills and project so to handle
static types.

The skills component requries a data.json file that should be placed in the `data` folder under app.

refactor(.gitignore, myWork/): ignore data.json that should be parse in on deployment.
Added the skills component into the `myWork` page.