refactor(components/skills.tsx, interfaces/skills_project.tsx, myWork/page.tsx): More or less completed skills section for myWork page
Skills card is now more or less completed with animation.
Skills will be written into the data.json then will be dynamically loading into the page.
On hover, the card will move slightly up and on click the card will flip to show a short description of what the skills was used for.

All interfaces in the skills_project.tsx is now export for ease of use and modularity.