refactor(.gitignore, imageProcess.ts, next.config.ts): Finished the auto downloaded of images for projects.
The auto downloaded of the image from the data.json is done in the `imageProcess.ts` file. It is called in next.config.ts. However the imageProcess.ts script is called multiple times in the next.config.ts so will need to fix this eventually. Added a new package `axios` for downloading images.